refactor: migrate media repository
