+++ server/src/types.ts
@@ -1,10 +1,11 @@
 import { UserEntity } from 'src/entities/user.entity';
-import { Permission } from 'src/enum';
+import { ExifOrientation, ImageFormat, Permission, TranscodeTarget, VideoCodec } from 'src/enum';
 import { AccessRepository } from 'src/repositories/access.repository';
 import { ActivityRepository } from 'src/repositories/activity.repository';
 import { ApiKeyRepository } from 'src/repositories/api-key.repository';
 import { AuditRepository } from 'src/repositories/audit.repository';
 import { ConfigRepository } from 'src/repositories/config.repository';
+import { MediaRepository } from 'src/repositories/media.repository';
 import { MemoryRepository } from 'src/repositories/memory.repository';
 import { ViewRepository } from 'src/repositories/view-repository';
 
@@ -24,6 +25,7 @@ export type IAccessRepository = { [K in keyof AccessRepository]: RepositoryInter
 export type IApiKeyRepository = RepositoryInterface<ApiKeyRepository>;
 export type IAuditRepository = RepositoryInterface<AuditRepository>;
 export type IConfigRepository = RepositoryInterface<ConfigRepository>;
+export type IMediaRepository = RepositoryInterface<MediaRepository>;
 export type IMemoryRepository = RepositoryInterface<MemoryRepository>;
 export type IViewRepository = RepositoryInterface<ViewRepository>;
 
@@ -39,3 +41,135 @@ export type ApiKeyItem =
 export type MemoryItem =
   | Awaited<ReturnType<IMemoryRepository['create']>>
   | Awaited<ReturnType<IMemoryRepository['search']>>[0];
+
+export interface CropOptions {
+  top: number;
+  left: number;
+  width: number;
+  height: number;
+}
+
+export interface ImageOptions {
+  format: ImageFormat;
+  quality: number;
+  size: number;
+}
+
+export interface RawImageInfo {
+  width: number;
+  height: number;
+  channels: 1 | 2 | 3 | 4;
+}
+
+interface DecodeImageOptions {
+  colorspace: string;
+  crop?: CropOptions;
+  processInvalidImages: boolean;
+  raw?: RawImageInfo;
+}
+
+export interface DecodeToBufferOptions extends DecodeImageOptions {
+  size: number;
+  orientation?: ExifOrientation;
+}
+
+export type GenerateThumbnailOptions = ImageOptions & DecodeImageOptions;
+
+export type GenerateThumbnailFromBufferOptions = GenerateThumbnailOptions & { raw: RawImageInfo };
+
+export type GenerateThumbhashOptions = DecodeImageOptions;
+
+export type GenerateThumbhashFromBufferOptions = GenerateThumbhashOptions & { raw: RawImageInfo };
+
+export interface GenerateThumbnailsOptions {
+  colorspace: string;
+  crop?: CropOptions;
+  preview?: ImageOptions;
+  processInvalidImages: boolean;
+  thumbhash?: boolean;
+  thumbnail?: ImageOptions;
+}
+
+export interface VideoStreamInfo {
+  index: number;
+  height: number;
+  width: number;
+  rotation: number;
+  codecName?: string;
+  frameCount: number;
+  isHDR: boolean;
+  bitrate: number;
+  pixelFormat: string;
+}
+
+export interface AudioStreamInfo {
+  index: number;
+  codecName?: string;
+  frameCount: number;
+}
+
+export interface VideoFormat {
+  formatName?: string;
+  formatLongName?: string;
+  duration: number;
+  bitrate: number;
+}
+
+export interface ImageDimensions {
+  width: number;
+  height: number;
+}
+
+export interface InputDimensions extends ImageDimensions {
+  inputPath: string;
+}
+
+export interface VideoInfo {
+  format: VideoFormat;
+  videoStreams: VideoStreamInfo[];
+  audioStreams: AudioStreamInfo[];
+}
+
+export interface TranscodeCommand {
+  inputOptions: string[];
+  outputOptions: string[];
+  twoPass: boolean;
+  progress: {
+    frameCount: number;
+    percentInterval: number;
+  };
+}
+
+export interface BitrateDistribution {
+  max: number;
+  target: number;
+  min: number;
+  unit: string;
+}
+
+export interface ImageBuffer {
+  data: Buffer;
+  info: RawImageInfo;
+}
+
+export interface VideoCodecSWConfig {
+  getCommand(
+    target: TranscodeTarget,
+    videoStream: VideoStreamInfo,
+    audioStream: AudioStreamInfo,
+    format?: VideoFormat,
+  ): TranscodeCommand;
+}
+
+export interface VideoCodecHWConfig extends VideoCodecSWConfig {
+  getSupportedCodecs(): Array<VideoCodec>;
+}
+
+export interface ProbeOptions {
+  countFrames: boolean;
+}
+
+export interface VideoInterfaces {
+  dri: string[];
+  mali: boolean;
+}
